Fix argument of doubleVectorToArray.

+++ src/jumpingalien/model/Utilities.java
@@ -68,7 +68,7 @@ public class Utilities {
 	 * @param v The Vector2D to convert.
 	 * @return An array of 2 elements where the first element is v.x and the second element is v.y.
 	 */
-	public static double[] doubleVectorToArray(Vector2D<Integer> v){
+	public static double[] doubleVectorToArray(Vector2D<Double> v){
 		double[] vs = {v.x, v.y};
 		return vs;
 	}
